fix: redis 커넥션 설정 변경
